% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{add_filters}
\alias{add_filters}
\title{add_filters}
\usage{
add_filters(VCF, ...)
}
\arguments{
\item{VCF}{a dataframe of variants processed by the pipeline.}
}
\value{
A dataframe of the VCF with TRUE/FALSE columns for each filter
}
\description{
A function to add columns for each of the filters with TRUE/FALSE if the variant passes/fails
}
\examples{

add_filters(VCF)

}
\keyword{VCF}
