% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{add_metadata}
\alias{add_metadata}
\title{add_metadata}
\usage{
add_metadata(object, metadata, ...)
}
\arguments{
\item{metadata}{dataframe of sample metadata}
}
\value{
object of class VariantHelper with metadata added
}
\description{
A function to add metadata to the object
}
\examples{

add_metadata(obj, metadata)
}
\keyword{VCF}
