% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{add_summary}
\alias{add_summary}
\title{add_summary}
\usage{
add_summary(object, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}

\item{stat}{the statistic to be used. Must exist in the dataframes for each caller (default: AF)}

\item{FUN}{function to use for the summary (default: mean)}
}
\value{
object of class VariantHelper
}
\description{
A function to add a summary of a statistic from the individual callers to the consensus dataframe
}
\examples{

add_summary(obj)
}
\keyword{VCF}
