% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{plot_sample_stat}
\alias{plot_sample_stat}
\alias{plot_sample_freq}
\alias{plot_gene_freq}
\alias{plot_gene_stat}
\alias{plot_annotation}
\title{ancillary_plots}
\usage{
plot_sample_stat(variants, stat = "VAF", category = Tumor_Sample_Barcode, ...)

plot_sample_freq(
  variants,
  category = Tumor_Sample_Barcode,
  Type = Variant_Classification,
  COLOUR = NULL,
  ...
)

plot_gene_freq(
  variants,
  Gene = Gene,
  category = Tumor_Sample_Barcode,
  Type = Variant_Classification,
  COLOUR = NULL,
  ...
)

plot_gene_stat(variants, stat = "VAF", Gene = Gene, ...)

plot_annotation(
  variants,
  full.variants,
  category = Tumor_Sample_Barcode,
  annotation_cols = NULL,
  annotation_colours = NULL,
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{stat}{character matching the numerical column to plot (default: "VAF")}

\item{category}{column name to use for sample ID (default: Tumor_Sample_Barcode)}

\item{Type}{column name to use for fill of plot (default: Variant_Classification)}

\item{COLOUR}{colours for the frequency plot (default: NULL)}

\item{Gene}{column name to use for gene ID (default: Gene)}

\item{full.variants}{dataframe of unfiltered variants to get annotation information from}

\item{annotation_cols}{character/vector names of columns to add to the annotation plot (default: NULL)}

\item{annotation_colours}{named vector for annotation plot colours}
}
\description{
A collection of functions to plot the sample/gene level plots
}
\examples{

plot_sample_stat(variants)


plot_sample_stat(variants)


plot_gene_freq(variants)


plot_gene_stat(variants)


plot_gene_stat(variants, annotation_cols = c("sampleType", "donorID"))

}
