% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{annotate_mutation_dist}
\alias{annotate_mutation_dist}
\alias{annotate_mutation_dist.data.frame}
\alias{annotate_mutation_dist.list}
\alias{annotate_mutation_dist.VariantHelper}
\title{annotate_mutation_dist}
\usage{
annotate_mutation_dist(x)

\method{annotate_mutation_dist}{data.frame}(mut_dist, Genes = NULL, ...)

\method{annotate_mutation_dist}{list}(list, ...)

\method{annotate_mutation_dist}{VariantHelper}(obj, ...)
}
\arguments{
\item{mut_dist}{data frame of mutation positions}

\item{Genes}{GRanges object of genes to annotate}
}
\description{
annotate_mutation_dist
}
\examples{

annotate_mutation_dist()

annotate_mutation_dist(df)

annotate_mutation_dist(list)

annotate_mutation_dist(obj)
}
\keyword{VCF}
