% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\name{assemble_upset}
\alias{assemble_upset}
\alias{assemble_upset.upsetti_spaghetti}
\title{assemble_upset}
\usage{
assemble_upset(x, ...)

\method{assemble_upset}{upsetti_spaghetti}(obj, ...)
}
\value{
an upsetti_spaghetti object
}
\description{
combine the plot
}
\examples{

assemble_upset(obj)
}
\keyword{VCF}
