% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot_processing.R
\name{calculate_design}
\alias{calculate_design}
\title{calculate_design}
\usage{
calculate_design(top, side, annot, main.scale = 6, ...)
}
\arguments{
\item{top}{the list of plots for the top of the plot}

\item{side}{the list of plots for the side of the plot}

\item{main.scale}{the relative size of the main plot (default: 6)}
}
\description{
Work out the design string from the given top and side plots
}
\examples{

calculate_design(top, side, , 10)

}
