% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{combine_calls}
\alias{combine_calls}
\title{combine_calls}
\usage{
combine_calls(object, ...)
}
\arguments{
\item{object}{object of class VariantHelper}
}
\value{
Long data frame of calls from all the callers.
}
\description{
A function to take data from selected callers and combine them into a long dataframe
}
\examples{

combine_calls(obj)

combine_calls(obj)
}
\keyword{VCF}
