% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{combine_objects}
\alias{combine_objects}
\title{combine_objects}
\usage{
combine_objects(obj, obj2)
}
\arguments{
\item{obj}{object of class VariantHelper}

\item{obj2}{object of class VariantHelper}
}
\value{
combined object of class VariantHelper
}
\description{
A function to combine two instances of VariantHelper
}
\examples{

combine_objects(obj, obj2)
}
\keyword{VCF}
