% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{count_vars_by_caller_sample}
\alias{count_vars_by_caller_sample}
\title{count_vars_by_caller_sample}
\usage{
count_vars_by_caller_sample(object, category = Sample, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}
}
\value{
dataframe of the number of times a variant was called
}
\description{
A function to count the number of callers that called each variant.
}
\examples{

count_vars_by_caller_sample(obj)
}
\keyword{VCF}
