% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{detect_kataegis}
\alias{detect_kataegis}
\alias{detect_kataegis.data.frame}
\alias{detect_kataegis.list}
\alias{detect_kataegis.VariantHelper}
\title{detect_kataegis}
\usage{
detect_kataegis(x)

\method{detect_kataegis}{data.frame}(df)

\method{detect_kataegis}{list}(list)

\method{detect_kataegis}{VariantHelper}(obj)
}
\arguments{
\item{df}{of mutations}

\item{list}{of mutation dataframes}

\item{obj}{of type VariantHelper}

\item{Genes}{GRanges object of genes to annotate}
}
\description{
detect_kataegis
}
\examples{

detect_kataegis(df)

detect_kataegis(list)

detect_kataegis(obj)
}
\keyword{VCF}
