% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{filter_by_consensus}
\alias{filter_by_consensus}
\title{filter_by_consensus}
\usage{
filter_by_consensus(object, consensus, category = Sample, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{consensus}{vector of Variant IDs for consensus variants}
}
\value{
dataframe of the variants that exist in consensus
}
\description{
A function to retain only the varians in consensus
}
\examples{

filter_by_consensus(obj)
}
\keyword{VCF}
