% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filter_variants}
\alias{filter_variants}
\title{update_stats}
\usage{
filter_variants(object, ...)
}
\arguments{
\item{biotype}{A vector of biotypes to select (default: c("protein_coding"))}

\item{impact}{A vector of biotypes to select (default: c("HIGH", "MODERATE"))}

\item{existing}{bool of whether to filter by dbSNP/COSMIC (default: TRUE)}

\item{population}{numeric value to filter 1000g/gnomAD (default: 0.01)}

\item{vaf}{numeric value to filter vaf (default: 0)}

\item{depth}{numeric value to filter depth (default: 1)}

\item{alt.depth}{numeric value to filter alt allele depth (default: 1)}
}
\value{
object of class VariantHelper with filters applied
}
\description{
A wrapper to filter all VCFs loaded in obj@Callers
}
\examples{

filter_variants(obj, vaf = 0.5)
}
\keyword{VCF}
