% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filter_vars}
\alias{filter_vars}
\title{filter_vars}
\usage{
filter_vars(
  VCF,
  returnStats = FALSE,
  returnPlot = FALSE,
  plotStats = FALSE,
  printStats = FALSE,
  ...
)
}
\arguments{
\item{VCF}{the VCF file process through the pipeline}

\item{returnStats}{return a filtering summary (default: FALSE)}

\item{returnPlot}{return a plot summary of the filtering (default: FALSE)}

\item{plotStats}{print a plot summary of the filtering (default: FALSE)}

\item{printStats}{print a summary of the filtering (default: FALSE)}

\item{biotype}{A vector of biotypes to select (default: c("protein_coding"))}

\item{impact}{A vector of biotypes to select (default: c("HIGH", "MODERATE"))}

\item{existing}{bool of wether to filter by dbSNP/COSMIC (default: FALSE)}

\item{population}{numeric value to filter 1000g/gnomAD (default: 0.01)}

\item{vaf}{numeric value to filter vaf (default: 0)}

\item{depth}{numeric value to filter depth (default: 1)}

\item{alt.depth}{numeric value to filter alt allele depth (default: 1)}

\item{returnDF}{bool return dataframe rather than vector (default: FALSE)}
}
\value{
A dataframe of the filtered VCF (or a plot/overview of the filtering)
}
\description{
A function to filter processed VCF calls
}
\examples{

filter_vars(VCF)
filter_vars(VCF, returnStats = FALSE)
filter_vars(VCF, af = 0.1)

}
\keyword{VCF}
