% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\name{generate_upsetti}
\alias{generate_upsetti}
\title{generate_upsetti}
\usage{
generate_upsetti(df, cols = NULL, ...)
}
\arguments{
\item{df}{a wide data.frame of data to plot}

\item{cols}{a vector of columns to pivot on}
}
\value{
a new upsetti_spaghetti object
}
\description{
A function to generate and process data into an upsetti_spaghetti class
}
\examples{

generate_upsetti(df, cols = c("col1", "col2"))
}
\keyword{VCF}
