% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\name{generate_X_plot}
\alias{generate_X_plot}
\alias{generate_intersect_plot}
\alias{generate_intersect_plot.data.frame}
\alias{generate_intersect_plot.upsetti_spaghetti}
\alias{generate_set_plot}
\alias{generate_set_plot.data.frame}
\alias{generate_set_plot.upsetti_spaghetti}
\alias{generate_group_plot}
\alias{generate_group_plot.data.frame}
\alias{generate_group_plot.upsetti_spaghetti}
\title{generate_X_plot}
\usage{
generate_intersect_plot(x, ...)

\method{generate_intersect_plot}{data.frame}(df, id, colour = NULL, col_vec = NULL, ...)

\method{generate_intersect_plot}{upsetti_spaghetti}(obj, ...)

generate_set_plot(x, ...)

\method{generate_set_plot}{data.frame}(df, colour = NULL, col_vec, ...)

\method{generate_set_plot}{upsetti_spaghetti}(obj, ...)

generate_group_plot(x, ...)

\method{generate_group_plot}{data.frame}(df, id, set_plot, intersect_plot, ...)

\method{generate_group_plot}{upsetti_spaghetti}(obj, ...)
}
\arguments{
\item{df}{a long data.frame of data to plot}

\item{id}{a column in the data for id-ing the individuals}

\item{colour}{a column to colour data by}

\item{col_vec}{a named vector of colours for the plot}

\item{set_plot}{plot for the side of the upset}

\item{intersect_plot}{plot for the top of the upset}
}
\value{
a plot of the intersection
}
\description{
functions to generate the parts of an upset plot
}
\examples{

generate_intersect_plot(df, id = col)
generate_set_plot(df, id = col)
generate_group_plot(df, id = col)
}
\keyword{VCF}
