% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{get_all_ens_genes}
\alias{get_all_ens_genes}
\title{get_all_ens_genes}
\usage{
get_all_ens_genes(mart = NULL)
}
\arguments{
\item{mart}{object of type mart}
}
\description{
get_all_ens_genes
}
\examples{

get_all_ens_genes()
}
\keyword{VCF}
