% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinimumReportingStandards.R
\name{get_min_reporting_stats}
\alias{get_min_reporting_stats}
\title{get_min_reporting_stats}
\usage{
get_min_reporting_stats(...)
}
\arguments{
\item{path}{path to QC folder}

\item{pattern}{pattern to find files (e.g. .con. to ignoire .raw.)}

\item{samtools_suffix}{folder with samtools output default: "/samtools_stats/"}

\item{picard_suffix}{folder with picard tools output default: /depth/picard/}
}
\value{
data frame of metrics (cols: Metrics, Value, Sample, Group)
}
\description{
Get the minimum reporting standards from picard and samtools
}
\examples{

get_min_reporting_stats("QC")
}
\keyword{VCF}
