% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinimumReportingStandards.R
\name{get_picard_stat}
\alias{get_picard_stat}
\title{get_picard_stat}
\usage{
get_picard_stat(path, picard_suffix = "/depth/picard/", ...)
}
\arguments{
\item{path}{path to QC folder}

\item{picard_suffix}{folder with picardtools output default: /depth/picard/}
}
\value{
data frame of metrics from picardtools (metrics:
MEAN_TARGET_COVERAGE, ZERO_CVG_TARGETS_PCT, PCT_TARGET_BASES_10X,
PCT_TARGET_BASES_30X, PCT_TARGET_BASES_100X;
cols:
Metrics, Value, Sample, Group)
}
\description{
Get minimum reporting stats from picardtools
}
\examples{

get_picard_stat("QC")
}
\keyword{VCF}
