% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinimumReportingStandards.R
\name{get_samtools_stat}
\alias{get_samtools_stat}
\title{get_samtools_stat}
\usage{
get_samtools_stat(path, pattern, samtools_suffix = "/samtools_stats/", ...)
}
\arguments{
\item{path}{path to QC folder}

\item{pattern}{pattern to find files (e.g. .con. to ignoire .raw.)}

\item{samtools_suffix}{folder with samtools output default: "/samtools_stats/"}
}
\value{
data frame of metrics from samtools (metrics:
raw total sequences, reads mapped, reads unmapped, reads properly paired,
reads duplicated, average length, insert size average, pairs on different chromosomes"
cols:
Metrics, Value, Sample, Group)
}
\description{
Get minimum reporting stats from samtools
}
\examples{

get_samtools_stat("QC", ".con.txt")
}
\keyword{VCF}
