% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentQC.R
\name{load_depth_qc}
\alias{load_depth_qc}
\title{filter_vars}
\usage{
load_depth_qc(
  folders = "QC/depth/",
  pattern = ".*sample_summary",
  meta.data = NULL,
  ...
)
}
\arguments{
\item{folders}{A vector/string of folders to look in for files (default: QC/depth/).}

\item{pattern}{a regex for finding the right files (default: .*sample_summary)}

\item{meta.data}{a dataframe containing aditional sample level information.}
}
\value{
A dataframe of the GATK depth summaries.
}
\description{
A function to load the depth information from the Okosun lab alignment snakemake pipelines (or anything that uses GATK DepthOfCoverage)
}
\examples{

load_depth_qc()

load_depth_qc(folders = "WES/QC/depth", meta.data = MData)

}
\keyword{VCF}
