% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot_processing.R
\name{logging}
\alias{logging}
\title{logging}
\usage{
logging(verbose = FALSE, text = "", ...)
}
\arguments{
\item{verbose}{bool should the message be printed}

\item{text}{the text to plot}
}
\description{
logging
}
