% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{set_multi_hit}
\alias{set_multi_hit}
\title{set_multi_hit}
\usage{
set_multi_hit(
  variants,
  category = Tumor_Sample_Barcode,
  Gene = Gene,
  Type = Variant_Classification,
  multi_hit_label = "multi_hit",
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{category}{column name to use for sample ID (default: Tumor_Sample_Barcode)}

\item{Gene}{column name to use for gene ID (default: Gene)}

\item{Type}{column name to use for fill of plot (default: Variant_Classification)}

\item{multi_hit_label}{string to use for multi hits (default: multi_hit)}

\item{stat}{character matching the numerical column to plot (default: "VAF")}

\item{MultiHit}{bool collapse multi hits to multihit (default: TRUE)}

\item{MultiHitSG}{bool collapse multi hits to multihit in this plot. If NULL defaults to MultiHit (default: NULL)}
}
\description{
A function to convert multiple hits on the same gene in the same sample to a single entry labelled "multi hit"
}
\examples{

set_multi_hit(variants)

}
