% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{mut_dist}
\alias{mut_dist}
\title{mut_dist}
\usage{
mut_dist(mutations, distance = 1000, ...)
}
\arguments{
\item{mutations}{data frame of mutation positions}

\item{distance}{max mean distance for variants (default: 1000)}
}
\description{
mut_dist
}
\examples{

mut_dist(mutations)
}
\keyword{VCF}
