% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{mutation_dist_to_grange}
\alias{mutation_dist_to_grange}
\alias{mutation_dist_to_grange.data.frame}
\alias{mutation_dist_to_grange.list}
\alias{mutation_dist_to_grange.VariantHelper}
\title{mutation_dist_to_grange}
\usage{
mutation_dist_to_grange(x)

\method{mutation_dist_to_grange}{data.frame}(df)

\method{mutation_dist_to_grange}{list}(list)

\method{mutation_dist_to_grange}{VariantHelper}(obj)
}
\arguments{
\item{df}{of mutations distances}

\item{list}{of mutation distances dataframes}

\item{obj}{of type VariantHelper}
}
\description{
mutation_dist_to_grange
}
\examples{

mutation_dist_to_grange()

mutation_dist_to_grange(df)

mutation_dist_to_grange(list)

mutation_dist_to_grange(obj)
}
\keyword{VCF}
