% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{onkoplot}
\alias{onkoplot}
\title{Onkoplot}
\usage{
onkoplot(variants, ...)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{stat}{character matching the numerical column to plot (default: "VAF")}

\item{category}{column name to use for sample ID (default: Tumor_Sample_Barcode)}

\item{Gene}{column name to use for gene ID (default: Gene)}

\item{Type}{column name to use for fill of plot (default: Variant_Classification)}

\item{frequency}{bool whether to plot the frequency}

\item{sample_frequency}{bool whether to plot the sample frequency. Overwritten by frequency if NULL (default: NULL)}

\item{gene_frequency}{bool whether to plot the sample frequency. Overwritten by frequency if NULL (default: NULL)}

\item{stats}{vector/character of the stats to plot (default: c("VAF"))}

\item{sample_stats}{vector/character of the sample level stats to plot. Overwritten by stats if NULL (default: NULL)}

\item{gene_stats}{vector/character of the sample level stats to plot. Overwritten by stats if NULL (default: NULL)}
}
\description{
Create an onkoplot
}
\examples{

plot_gene_stat(variants)

}
