% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentQC.R
\name{plot_depth_qc}
\alias{plot_depth_qc}
\title{filter_vars}
\usage{
plot_depth_qc(depth = NULL, category = Sample, ...)
}
\arguments{
\item{depth}{dataframe of depth of coverage statistics. If NULL will try to load data using load_depth_qc (default: NULL)}

\item{category}{column in depth to use for plotting}
}
\value{
A ggplot object
}
\description{
A function to plot the depth information
}
\examples{

plot_depth_qc(depth)

}
\keyword{VCF}
