% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{plot_filters}
\alias{plot_filters}
\title{plot_filters}
\usage{
plot_filters(Stats, title = "", ...)
}
\arguments{
\item{Stats}{The boolean filter columns from the VCF}

\item{title}{The title for the plot (default: "")}
}
\value{
two barplots of the variants that pass each filter and the number of filters passed
}
\description{
A function to plot the summary of the variants that pass the filters
}
\examples{

plot_filters(VCF)

}
\keyword{VCF}
