% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{plot_gene_sample_mutations}
\alias{plot_gene_sample_mutations}
\title{plot_gene_sample_mutations}
\usage{
plot_gene_sample_mutations(
  variants,
  category = Tumor_Sample_Barcode,
  MultiHit = TRUE,
  MultiHitSG = NULL,
  Gene = Gene,
  Type = Variant_Classification,
  COLOUR = NULL,
  xlab = "Sample",
  xnames = FALSE,
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{category}{column name to use for sample ID (default: Tumor_Sample_Barcode)}

\item{MultiHit}{bool collapse multi hits to multihit (default: TRUE)}

\item{MultiHitSG}{bool collapse multi hits to multihit in this plot. If NULL defaults to MultiHit (default: NULL)}

\item{Gene}{column name to use for gene ID (default: Gene)}

\item{Type}{column name to use for fill of plot (default: Variant_Classification)}

\item{stat}{character matching the numerical column to plot (default: "VAF")}

\item{MultiHitS}{bool collapse multi hits to multihit in the sample plot. If NULL defaults to MultiHit (default: NULL)}

\item{MultiHitG}{bool collapse multi hits to multihit in the gene plot. If NULL defaults to MultiHit (default: NULL)}
}
\description{
A function to generate a tile plot of the variants in each sample.
}
\examples{

plot_gene_stat(variants)

}
