% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{plot_recurrent_vars}
\alias{plot_recurrent_vars}
\title{plot_recurrent_vars}
\usage{
plot_recurrent_vars(obj = NULL, df = NULL, ...)
}
\arguments{
\item{obj}{an object of type VariantHelper (default: NULL)}
}
\value{
ggplot of recurrent variant positions
}
\description{
Plot out variants that occur more than once
}
\examples{

return_recurrent_vars(obj)
}
\keyword{VCF}
