% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\name{plot_upset}
\alias{plot_upset}
\title{plot_upset}
\usage{
plot_upset(df, ...)
}
\value{
an upsetti_spaghetti object
}
\description{
Take a data.frame and return an upset plot stored in an object of the class upsetti_spaghetti
}
\examples{

plot_upset(obj)
}
\keyword{VCF}
