% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_AlleleMix}
\alias{process_AlleleMix}
\title{process_AlleleMix}
\usage{
process_AlleleMix(row, RA = "ALT", ...)
}
\arguments{
\item{row}{A row of a VCF file}

\item{RA}{The allele to return (REF/ALT) Default = "ALT")}
}
\value{
A dataframe of the calls with the VEP annotations separated into columns. Multiple transcripts will be split into different columns.
}
\description{
A function to split VEP annotations into columns
}
\examples{

process_AlleleMix(row, RA = "ALT")
}
\keyword{VCF}
