% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_counts}
\alias{process_counts}
\title{process_counts}
\usage{
process_counts(file, VCF, Caller = "", ...)
}
\arguments{
\item{file}{The path to the file to be read}

\item{VCF}{the VCF file with the tumour format split}

\item{Caller}{the caller the data is from. If empty determines from filepath (default: "")}

\item{verbose}{print out extra text for debugging (verbose: FALSE)}
}
\value{
A dataframe of the VCF with ADR, RDR, DP and AF columns
}
\description{
A function to process read counts from VCF files
}
\examples{

process_counts(file, VCF)
}
\keyword{VCF}
