% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{process_mutation_distances}
\alias{process_mutation_distances}
\alias{process_mutation_distances.data.frame}
\alias{process_mutation_distances.list}
\alias{process_mutation_distances.VariantHelper}
\title{process_mutation_distances}
\usage{
process_mutation_distances(x)

\method{process_mutation_distances}{data.frame}(mutations, ...)

\method{process_mutation_distances}{list}(list, ...)

\method{process_mutation_distances}{VariantHelper}(obj, ...)
}
\arguments{
\item{mutations}{data frame of mutation positions}

\item{obj}{object of class variant helpers}
}
\description{
process_mutation_distances
}
\examples{

process_mutation_distances()

process_mutation_distances(df)

process_mutation_distances(list)

process_mutation_distances(obj)
}
\keyword{VCF}
