% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_Mutect2}
\alias{process_Mutect2}
\title{process_Mutect2}
\usage{
process_Mutect2(VCF, ...)
}
\arguments{
\item{VCF}{a dataframe of VCF records from Mutect2 with the TUMOR columns separated}
}
\value{
A dataframe of the calls with the VEP annotations separated into columns. Multiple transcripts will be split into different columns.
}
\description{
A function to split VEP annotations into columns
}
\examples{

process_Mutect2(row)
}
\keyword{VCF}
