% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_Strelka2}
\alias{process_Strelka2}
\title{process_Strelka2}
\usage{
process_Strelka2(VCF, ...)
}
\arguments{
\item{VCF}{a dataframe of VCF records from Strelka with the TUMOR columns separated}
}
\value{
A dataframe of the calls with the VEP annotations separated into columns. Multiple transcripts will be split into different columns.
}
\description{
A function to split VEP annotations into columns
}
\examples{

process_Strelka2(row)
}
\keyword{VCF}
