% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\name{process_upsetti_plots}
\alias{process_upsetti_plots}
\title{process_upsetti_plots}
\usage{
process_upsetti_plots(obj, ...)
}
\arguments{
\item{an}{upsetti_spaghetti object}
}
\value{
an upsetti_spaghetti object with all 3 plots generated
}
\description{
combine the plot
}
\examples{

assemble_upset(obj)
}
\keyword{VCF}
