% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_VarDict}
\alias{process_VarDict}
\title{process_VarDict}
\usage{
process_VarDict(VCF, tumourOnly = FALSE, ...)
}
\arguments{
\item{VCF}{a dataframe of VCF records from VarDict with the TUMOR columns separated}
}
\value{
A dataframe of the calls with the VEP annotations separated into columns. Multiple transcripts will be split into different columns.
}
\description{
A function to split VEP annotations into columns
}
\examples{

process_VarDict(row)
}
\keyword{VCF}
