% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_Varscan2}
\alias{process_Varscan2}
\title{process_Varscan2}
\usage{
process_Varscan2(VCF, sample = NULL, tumourOnly = FALSE, ...)
}
\arguments{
\item{VCF}{a dataframe of VCF records from Varscan2 with the TUMOR columns separated}
}
\value{
A dataframe of the calls with the VEP annotations separated into columns. Multiple transcripts will be split into different columns.
}
\description{
A function to split VEP annotations into columns
}
\examples{

process_Varscan2(row)
}
\keyword{VCF}
