% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_vcf}
\alias{process_vcf}
\title{read_VCF}
\usage{
process_vcf(file, ...)
}
\arguments{
\item{file}{The path to the file to be read}
}
\value{
A dataframe of the VCF skipping the commented lines
}
\description{
A function read a VCF file removing the header (denoted by the ##)
}
\examples{

read_VCF(file)
}
\keyword{VCF}
