% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{raise_sample_pattern_error}
\alias{raise_sample_pattern_error}
\title{split_format}
\usage{
raise_sample_pattern_error(colnames, pattern)
}
\arguments{
\item{colnames}{column names for error message}

\item{pattern}{pattern for error message}
}
\value{
Error
}
\description{
A function to give an error for sample column finding
}
\keyword{VCF}
