% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{read_VCF}
\alias{read_VCF}
\title{read_VCF}
\usage{
read_VCF(file, pass = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{file}{The path to the file to be read}

\item{pass}{whether to filter out non PASS variants (default: TRUE)}

\item{verbose}{print out extra text for debugging (verbose: FALSE)}

\item{pick}{whether to filter out the annotations that weren't picked (default: TRUE)}
}
\value{
A dataframe of the VCF skipping the commented lines
}
\description{
A function to read a VCF file whilst removing the header (denoted by the ##)
}
\examples{

read_VCF(file)
}
\keyword{VCF}
