% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{return_callers}
\alias{return_callers}
\title{return_callers}
\usage{
return_callers(object, callers = NULL, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{callers}{vector of the names of variant callers stored in the Callers slot.}
}
\value{
List of dataframes containing variant call information from all/selected callers.
}
\description{
A function to return the data frames holding variant calls.
Can optionally return a subset of the callers defined by a vector
}
\examples{

return_callers(obj, metadata)
}
\keyword{VCF}
