% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{return_cluster_format}
\alias{return_cluster_format}
\title{return_cluster_format}
\usage{
return_cluster_format(df, idxs, ...)
}
\arguments{
\item{df}{data frame of mutation positions}

\item{idxs}{indexes to return}
}
\description{
return_cluster_format
}
\examples{

return_cluster_format(df, idx)
}
\keyword{VCF}
