% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{return_consensus}
\alias{return_consensus}
\title{return_consensus}
\usage{
return_consensus(object)
}
\arguments{
\item{object}{object of class VariantHelper}
}
\value{
dataframe of consensus calls with the data in meta.data attached.
}
\description{
A function return the consensus calls from the object with the meta.data added.
}
\examples{

return_consensus(obj)
}
\keyword{VCF}
