% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kataegis.R
\name{return_mean_pos_difference}
\alias{return_mean_pos_difference}
\title{return_mean_pos_difference}
\usage{
return_mean_pos_difference(df, ...)
}
\arguments{
\item{df}{data frame of mutation positions}
}
\description{
return_mean_pos_difference
}
\examples{

return_mean_pos_difference(df, idx)
}
\keyword{VCF}
