% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{return_recurrent_vars}
\alias{return_recurrent_vars}
\title{return_recurrent_vars}
\usage{
return_recurrent_vars(obj, limit = 1, ...)
}
\arguments{
\item{obj}{an object of type VariantHelper}

\item{limit}{threshold for calling a variant recurrent (default: 1)}
}
\value{
data frame of variants and the number of times they are called.
}
\description{
Find variant positions that occur across several samples
}
\examples{

return_recurrent_vars(obj)
}
\keyword{VCF}
