% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{sample_plot}
\alias{sample_plot}
\alias{gene_plot}
\title{sample_gene_plots}
\usage{
sample_plot(
  variants,
  full.variants,
  frequency = TRUE,
  sample_frequency = NULL,
  stats = c("VAF"),
  sample_stats = NULL,
  MultiHitS = FALSE,
  MultiHit = TRUE,
  filter.vars = FALSE,
  ...
)

gene_plot(
  variants,
  frequency = TRUE,
  gene_frequency = NULL,
  stats = c("VAF"),
  gene_stats = NULL,
  MultiHitG = NULL,
  MultiHit = TRUE,
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{full.variants}{dataframe of all variant calls}

\item{frequency}{bool whether to plot the frequency}

\item{sample_frequency}{bool whether to plot the sample frequency. Overwritten by frequency if NULL (default: NULL)}

\item{stats}{vector/character of the stats to plot (default: c("VAF"))}

\item{sample_stats}{vector/character of the sample level stats to plot. Overwritten by stats if NULL (default: NULL)}

\item{filter.vars}{whether to filter to mutations in just the sample of interest.}

\item{gene_frequency}{bool whether to plot the sample frequency. Overwritten by frequency if NULL (default: NULL)}

\item{gene_stats}{vector/character of the sample level stats to plot. Overwritten by stats if NULL (default: NULL)}
}
\description{
A collection of functions lists of the top or side plots.
}
\examples{

plot_gene_stat(variants)

}
