% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot_processing.R
\name{get_weights}
\alias{get_weights}
\alias{get_n_top_genes}
\alias{order_sample_by_weight}
\alias{set_order}
\alias{order_samples_and_genes}
\title{get_weights}
\usage{
get_weights(number, ...)

get_n_top_genes(
  vars,
  number = 10,
  category = Tumor_Sample_Barcode,
  Gene = Gene,
  collapseCategory = TRUE,
  ...
)

order_sample_by_weight(
  variants,
  genes,
  category = Tumor_Sample_Barcode,
  Gene = Gene,
  reorder_by = NULL,
  ...
)

set_order(
  variants,
  genes,
  samples,
  category = Tumor_Sample_Barcode,
  Gene = Gene,
  reorder_by = NULL,
  ...
)

order_samples_and_genes(variants, ...)
}
\arguments{
\item{number}{the maximum number of genes to return}

\item{category}{column to use to stratify samples (default: Tumor_Sample_Barcode)}

\item{Gene}{column to use to stratify genes (default: Gene)}

\item{collapseCategory}{whether to collapse the counts at a sample level, i.e. number of genes mutated one time OR MORE per samples (default: TRUE)}

\item{variants}{dataframe of variants}

\item{genes}{dataframe of genes with gene name, number of occurances and weight}
}
\description{
weight numbers to allow cumulative ordering

get the most mutated genes

weight numbers to allow cumulative ordering

weight numbers to allow cumulative ordering

weight numbers to allow cumulative ordering
}
\examples{

get_weights(variants, 10)


get_weights(variants, 10)


get_weights(variants, 10)


get_weights(variants, 10)


get_weights(variants, 10)

}
