% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{set_colour_palette}
\alias{set_colour_palette}
\title{set_colour_palette}
\usage{
set_colour_palette(
  variants,
  colour = NULL,
  Type = Variant_Classification,
  multi_hit_label = "multi_hit",
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{colour}{vector of colours for the fill}

\item{Type}{column name to use for fill of plot (default: Variant_Classification)}
}
\description{
define the colour palette
}
\examples{

(variants)

}
