% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{set_sample_name}
\alias{set_sample_name}
\title{set_sample_name}
\usage{
set_sample_name(
  VCF,
  filename,
  sample = NULL,
  tumourPattern = NULL,
  normalPattern = NULL,
  ...
)
}
\arguments{
\item{VCF}{the VCF file as a dataframe}

\item{filename}{the name of the VCF file}

\item{sample}{the sample ID. If NULL this will be determined by the filename (default: NULL)}

\item{tumourPattern}{The pattern for the tumour sample ids. Used to rename the tumour sample column to tumour (default: NULL)}

\item{normalPattern}{The pattern for the normal sample ids. Used to rename the normal sample column to NORMAL (default: NULL)}
}
\value{
A dataframe of the tumour counts as columns
}
\description{
A function to determine the sample name and change to columns for the tumour and normal data to TUMOR and NORMAL if they aren't already
}
\examples{

set_sample_name(VCF, filename, tumourPattern = "tumour", normalPattern = "normal")
}
\keyword{VCF}
