% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot.R
\name{set_sample_order_full}
\alias{set_sample_order_full}
\title{set_sample_order_full}
\usage{
set_sample_order_full(
  variants,
  full.variants,
  category = Tumor_Sample_Barcode,
  ...
)
}
\arguments{
\item{variants}{dataframe of variant calls}

\item{full.variants}{dataframe of variant calls}

\item{category}{the column to use for the x axis}
}
\description{
set the sample order for the full variant list
}
\examples{

set_sample_order_full(variants, full.variants)

}
