% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{split_format}
\alias{split_format}
\title{split_format}
\usage{
split_format(VCF, ...)
}
\arguments{
\item{VCF}{the VCF file as a dataframe}
}
\value{
A dataframe of the tumour counts as columns
}
\description{
A function to split the tumour data from a VCF into columns
}
\examples{

split_format(file, VCF)
}
\keyword{VCF}
