% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{stats_by_caller}
\alias{stats_by_caller}
\title{stats_by_caller}
\usage{
stats_by_caller(object, Stat = AF, category = Sample, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}
}
\value{
ggplot of the chose statistic of variants.
}
\description{
A function to plot the chosen statistic for each variant per sample (or other metadata column) per caller.
}
\examples{

stats_by_caller(obj)
}
\keyword{VCF}
