% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{summarise_filters}
\alias{summarise_filters}
\title{summarise_filters}
\usage{
summarise_filters(Stats, ...)
}
\arguments{
\item{Stats}{The boolean filter columns from the VCF}
}
\value{
A list of the variants that pass each filter and how many variants pass how many filters
}
\description{
A function summarise the variants that pass the filters
}
\examples{

summarise_filters(VCF)

}
\keyword{VCF}
