% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{test}
\alias{test}
\title{test}
\usage{
test()
}
\arguments{
\item{file}{The path to the file to be read}
}
\value{
A dataframe of the VCF skipping the commented lines
}
\description{
A function fully process the VCF through the pipeline.
}
\examples{

test(file)
}
\keyword{VCF}
