% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{update_stats}
\alias{update_stats}
\title{update_stats}
\usage{
update_stats(object, ...)
}
\value{
object of class VariantHelper with the stats updated
}
\description{
A function to update the stats held by the object
}
\examples{

update_stats(obj)
}
\keyword{VCF}
