% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetti_spaghetti.R
\docType{class}
\name{upsetti_spaghetti}
\alias{upsetti_spaghetti}
\alias{filter.upsetti_spaghetti}
\alias{mutate.upsetti_spaghetti}
\title{upsetti_spaghetti}
\usage{
\method{filter}{upsetti_spaghetti}(object, ...)

\method{mutate}{upsetti_spaghetti}(object, ...)
}
\description{
upsetti_spaghetti
}
\details{
This class will hold upset data
}
\section{Slots}{

\describe{
\item{\code{plots}}{a list of plot objects}

\item{\code{data}}{a data.frame of the long form data}
}}

\examples{

upsetti_spaghetti()
}
\keyword{VCF}
\keyword{upset}
\keyword{variants}
