% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{variant_helper_from_folder_paths}
\alias{variant_helper_from_folder_paths}
\title{variant_helper_from_folder_paths}
\usage{
variant_helper_from_folder_paths(paths, ...)
}
\arguments{
\item{paths}{a vector of paths to folders containing VCF files}
}
\value{
object of class VariantHelper with the VCF files loaded
}
\description{
A function to generate an object of class VariantHelper for a vector of paths to folders
}
\examples{

variant_helper_from_folder_paths(Paths)
}
\keyword{VCF}
