% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{variant_helper_from_list_of_calls}
\alias{variant_helper_from_list_of_calls}
\title{variant_helper_from_list_of_calls}
\usage{
variant_helper_from_list_of_calls(Calls, meta.data = NULL, ...)
}
\arguments{
\item{Calls}{a list of processed VCF files}

\item{meta.data}{dataframe of metadata for samples (default: NULL)}
}
\value{
object of class VariantHelper with the VCF info from Calls
}
\description{
A function to generate an object of class VariantHelper for a vector of paths to folders
}
\examples{

variant_helper_from_list_of_calls(Calls)
}
\keyword{VCF}
