% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_helper_to_maf.R
\name{variant_helper_to_maf}
\alias{variant_helper_to_maf}
\title{variant_helper_to_maf}
\usage{
variant_helper_to_maf(obj)
}
\arguments{
\item{obj}{of type VariantHelper}
}
\description{
variant_helper_to_maf
}
\examples{

variant_helper_to_maf(obj)

}
\keyword{VCF}
