% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{variant_stat_summary}
\alias{variant_stat_summary}
\alias{stat_summary}
\title{variant_stat_summary}
\usage{
variant_stat_summary(object, category = Sample, stat = AF, FUN = mean, ...)

stat_summary(object, category = Sample, stat = AF, FUN = mean, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}

\item{stat}{the statistic to be used. Must exist in the dataframes for each caller (default: AF)}

\item{FUN}{function to use for the summary (default: mean)}

\item{min_caller}{minimum number of callers needed to retain variant (default: 3)}
}
\value{
A dataframe of the consensus variant calls with the stat summary attached.
}
\description{
A function to add variants called by n callers or more to the consensus slot of the object
}
\examples{

variant_stat_summary(obj)
}
\keyword{VCF}
