% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\docType{class}
\name{VariantHelper-class}
\alias{VariantHelper-class}
\alias{VariantHelper}
\title{VariantHelper Class}
\description{
Hold variant call data
}
\details{
This class will hold your vcf data
}
\section{Slots}{

\describe{
\item{\code{Callers}}{list of dataframes containing call data}

\item{\code{Samples}}{vecotr of sample names}

\item{\code{Consensus}}{dataframe of consensus calls}

\item{\code{CallIDs}}{vector}

\item{\code{meta.data}}{dataframe of metadata}

\item{\code{misc}}{list}

\item{\code{info}}{list}
}}

\examples{

VariantHelper()
}
\keyword{VCF}
\keyword{variants}
