% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onkoplot_processing.R
\name{filter_variants_by_genes}
\alias{filter_variants_by_genes}
\title{VariantProcessing}
\usage{
filter_variants_by_genes(variants, genes, Gene = Gene, ...)
}
\arguments{
\item{number}{the maximum number of items to weight}
}
\description{
weight numbers to allow cumulative ordering
}
\examples{

get_weights(variants, 10)

}
