% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\name{variants_by_caller}
\alias{variants_by_caller}
\title{variants_by_caller}
\usage{
variants_by_caller(object, category = Sample, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}
}
\value{
ggplot of the number of variants.
}
\description{
A function to plot the number of variants caller per sample (or other metadata column) per caller.
}
\examples{

variants_by_caller(obj)
}
\keyword{VCF}
