% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{add_consensus}
\alias{add_consensus}
\title{add_consensus}
\usage{
add_consensus(object, min_caller = 3, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{min_caller}{minimum number of callers needed to retain variant (default: 3)}
}
\value{
object of class VariantHelper with the consensus calls in the consensus slot.
}
\description{
A function to add variants called by n callers or more to the consensus slot of the object
}
\examples{

add_consensus(obj)
}
\keyword{VCF}
