% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{count_vars_by_caller_sample}
\alias{count_vars_by_caller_sample}
\title{count_vars_by_caller_sample}
\usage{
count_vars_by_caller_sample(object, category = Sample, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}
}
\value{
dataframe of the number of times a variant was called
}
\description{
#' A function to generate an upset plot of the consensus variants.
#'
#' @title upset_by_caller
#' @return upset plot of the variants called by the callers.
#' @param object object of class VariantHelper
#' @param category name of a column in the metadata (default: Sample)
#' @param colours a vector of colours to label the plot by with
#' @keywords VCF
#' @import ComplexUpset
#' @export
#' @examples
#'
#' upset_by_caller(obj)
}
\details{
upset_by_caller <- function(object,
                            set=Caller,
                            id = sam_var_id,
                            colour = Sample,
                            col_vec = NULL, ...) {
    combine_calls(object) %>%
        mutate(sam_var_id = paste(varID, Sample)) %>%
        filter(!is.na(Caller)) %>%
        plot_upset(set = {{set}},
                   id = {{id}},
                   colour = {{colour}},
                   col_vec = col_vec,
                   ...)
}
A function to count the number of callers that called each variant.
}
\examples{

count_vars_by_caller_sample(obj)
}
\keyword{VCF}
