% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_folder}
\alias{process_folder}
\title{process_folder}
\usage{
process_folder(folder, pattern = ".annotated.vcf", ...)
}
\arguments{
\item{folder}{String containing the path to a folder of VCF files.}
}
\value{
A dataframe of the VCF skipping the commented lines
}
\description{
A function read a VCF file removing the header (denoted by the ##)
}
\examples{

process_folder(file)
}
\keyword{VCF}
