% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingVCF.R
\name{process_VCFs}
\alias{process_VCFs}
\title{process_folder}
\usage{
process_VCFs(VCFs)
}
\arguments{
\item{file}{The path to the file to be read}
}
\value{
A dataframe of the VCF skipping the commented lines
}
\description{
A function read a VCF file removing the header (denoted by the ##)
}
\examples{

process_VCFs(file)
}
\keyword{VCF}
