% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusCalling.R
\name{upset_by_caller}
\alias{upset_by_caller}
\title{upset_by_caller}
\usage{
upset_by_caller(object, category = Sample, colours = NULL, ...)
}
\arguments{
\item{object}{object of class VariantHelper}

\item{category}{name of a column in the metadata (default: Sample)}

\item{colours}{a vector of colours to label the plot by with}
}
\value{
upset plot of the variants called by the callers.
}
\description{
A function to generate an upset plot of the consensus variants.
}
\examples{

upset_by_caller(obj)
}
\keyword{VCF}
