% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantHelperClass.R
\docType{class}
\name{VariantHelper-class}
\alias{VariantHelper-class}
\alias{VariantHelper}
\title{VariantHelper Class}
\description{
Hold variant call data
}
\details{
This class will hold your vcf data
}
\examples{
Community()
}
\keyword{VCF}
\keyword{variants}
